requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
pydantic>=2.0.0
python-dotenv>=0.19.0
graphviz>=0.20.1
